pyproject.toml
[tool.poetry]
name = "pulse"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
scrapy = "^2.11.2"
scrapy-playwright = "^0.0.41"
pydantic = "^2.9.2"
fake-useragent = "^1.5.1"
anthropic = "^0.36.2"
voyageai = "^0.2.4"
pinecone-client = {extras = ["grpc"], version = "^5.0.1"}
alembic = "^1.13.3"
sqlalchemy = "^2.0.36"
asyncpg = "^0.29.0"
dateparser = "^1.2.0"


[tool.poetry.group.dev.dependencies]
black = "^24.10.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
